extract project metadata logic into dedicated hook
