+++ src/hooks/useProjectImport.ts
@@ -1,6 +1,6 @@
-import { useCallback, useEffect, useMemo, useState, type SyntheticEvent } from "react"
-import { devLog } from "@/lib/logger"
+import { useCallback, useEffect, useMemo, useState } from "react"
 import type { ProjectInsightsPayload, ProjectMetricId, RepositoryLike } from "@/components/dashboard/projects/types"
+import { useProjectMetadata } from "@/hooks/useProjectMetadata"
 
 interface UseProjectImportParams {
   open: boolean
@@ -11,153 +11,19 @@ interface UseProjectImportParams {
   onOpenChange: (open: boolean) => void
 }
 
-interface MetadataPayload {
-  name?: string
-  title?: string
-  description?: string
-  favicon?: string | null
-  ogImage?: string | null
-  logo?: string | null
-}
-
-interface ExtractMetadataResponse {
-  projectData?: MetadataPayload
-  metadata?: MetadataPayload
-}
-
-interface ResolvedImage {
-  displayUrl: string
-  directUrl: string
-}
-
 const GITHUB_DEFAULT_DESCRIPTION_PATTERN = /^Contribute to .* development by creating an account on GitHub\.?$/i
-
-function generateGitHubOgImage(owner: string, repoName: string): { proxyUrl: string; directUrl: string } {
-  const directUrl = `https://opengraph.githubassets.com/${owner}/${repoName}`
-  const proxyUrl = `/api/portfolio/github-og-image?owner=${encodeURIComponent(owner)}&repo=${encodeURIComponent(repoName)}`
-  return { proxyUrl, directUrl }
-}
-
-function parseGitHubRepoFromUrl(url: string): { owner: string; repoName: string } | null {
-  try {
-    const parsedUrl = new URL(url)
-    if (parsedUrl.hostname !== "github.com") return null
-
-    const pathParts = parsedUrl.pathname.split("/").filter(Boolean)
-    if (pathParts.length < 2) return null
-
-    return { owner: pathParts[0], repoName: pathParts[1] }
-  } catch {
-    return null
-  }
-}
 
 function isGitHubUrl(url: string): boolean {
   return url.includes("github.com")
 }
 
-function isSvgDataUri(url: string): boolean {
-  return url.startsWith("data:image/svg") || url.startsWith("data:image/svg+xml")
-}
-
-function isLongDataUri(url: string): boolean {
-  return url.startsWith("data:") && url.length > 50000
-}
-
-function shouldUseProxy(url: string): boolean {
-  if (!url || url.startsWith("data:") || url.startsWith("blob:")) {
-    return false
-  }
-
-  try {
-    const parsedUrl = new URL(url)
-    const currentHostname = typeof window !== "undefined" ? window.location.hostname : ""
-    const isExternal =
-      !parsedUrl.hostname.includes("localhost") &&
-      !parsedUrl.hostname.includes("127.0.0.1") &&
-      parsedUrl.hostname !== currentHostname
-    return isExternal
-  } catch {
-    return false
-  }
-}
-
 function parseNumberValue(value: string): number | null {
   if (!value.trim()) return null
+
   const parsed = Number(value)
   if (!Number.isFinite(parsed) || parsed < 0) return null
+
   return Math.round(parsed)
-}
-
-function readFileAsDataUrl(file: File): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader()
-    reader.onload = () => resolve(String(reader.result ?? ""))
-    reader.onerror = () => reject(reader.error)
-    reader.readAsDataURL(file)
-  })
-}
-
-function getMetadataImage(metadata?: MetadataPayload, projectData?: MetadataPayload): ResolvedImage | null {
-  const candidates = [metadata?.ogImage, metadata?.logo, projectData?.logo]
-
-  for (const rawCandidate of candidates) {
-    if (!rawCandidate) continue
-
-    const candidate = rawCandidate.trim()
-    if (!candidate) continue
-    if (isSvgDataUri(candidate)) continue
-    if (isLongDataUri(candidate)) continue
-
-    if (shouldUseProxy(candidate)) {
-      return {
-        displayUrl: `/api/portfolio/og-image-proxy?url=${encodeURIComponent(candidate)}`,
-        directUrl: candidate,
-      }
-    }
-
-    return {
-      displayUrl: candidate,
-      directUrl: candidate,
-    }
-  }
-
-  return null
-}
-
-function getGitHubImageFromRepository(repo: RepositoryLike): { proxyUrl: string | null; directUrl: string | null } {
-  if (repo.githubOgImage) {
-    try {
-      const parsedUrl = new URL(repo.githubOgImage)
-      if (parsedUrl.hostname === "opengraph.githubassets.com") {
-        const pathParts = parsedUrl.pathname.split("/").filter(Boolean)
-        if (pathParts.length >= 2) {
-          const generated = generateGitHubOgImage(pathParts[0], pathParts[1])
-          return { proxyUrl: generated.proxyUrl, directUrl: repo.githubOgImage }
-        }
-      }
-    } catch {
-      // Ignore and continue with next strategy.
-    }
-  }
-
-  if (repo.fullName) {
-    const [owner, repoName] = repo.fullName.split("/")
-    if (owner && repoName) {
-      const generated = generateGitHubOgImage(owner, repoName)
-      return { proxyUrl: generated.proxyUrl, directUrl: generated.directUrl }
-    }
-  }
-
-  if (repo.htmlUrl && !repo.isImported) {
-    const parsed = parseGitHubRepoFromUrl(repo.htmlUrl)
-    if (parsed) {
-      const generated = generateGitHubOgImage(parsed.owner, parsed.repoName)
-      return { proxyUrl: generated.proxyUrl, directUrl: generated.directUrl }
-    }
-  }
-
-  return { proxyUrl: null, directUrl: null }
 }
 
 export function useProjectImport({
@@ -169,15 +35,8 @@ export function useProjectImport({
   onOpenChange,
 }: UseProjectImportParams) {
   const [projectUrl, setProjectUrl] = useState("")
-  const [isLoading, setIsLoading] = useState(false)
   const [title, setTitle] = useState("")
   const [description, setDescription] = useState("")
-  const [favicon, setFavicon] = useState<string | null>(null)
-  const [logo, setLogo] = useState<string | null>(null)
-  const [directLogoUrl, setDirectLogoUrl] = useState<string | null>(null)
-  const [isFromGitHubSelection, setIsFromGitHubSelection] = useState(false)
-  const [imageLoadError, setImageLoadError] = useState(false)
-  const [imageLoading, setImageLoading] = useState(false)
   const [categories, setCategories] = useState<string[]>([])
   const [status, setStatus] = useState("")
   const [revenueInput, setRevenueInput] = useState("")
@@ -186,6 +45,30 @@ export function useProjectImport({
   const [technologies, setTechnologies] = useState<string[]>([])
   const [activeMetric, setActiveMetric] = useState<ProjectMetricId | null>(null)
   const [githubSearchQuery, setGithubSearchQuery] = useState("")
+
+  const projectMetadata = useProjectMetadata({
+    open,
+    projectUrl,
+    setTitle,
+    setDescription,
+  })
+
+  const {
+    isLoading,
+    favicon,
+    logo,
+    directLogoUrl,
+    imageLoadError,
+    imageLoading,
+    resetMetadata,
+    setIsFromGitHubSelection,
+    applyGitHubRepository,
+    handleUploadImage,
+    handleImageLoadStart,
+    handleImageLoad,
+    handleImageError,
+    captureScreenshot,
+  } = projectMetadata
 
   const filteredRepos = useMemo(() => {
     const searchText = githubSearchQuery.trim().toLowerCase()
@@ -207,12 +90,6 @@ export function useProjectImport({
     setProjectUrl("")
     setTitle("")
     setDescription("")
-    setFavicon(null)
-    setLogo(null)
-    setDirectLogoUrl(null)
-    setIsFromGitHubSelection(false)
-    setImageLoadError(false)
-    setImageLoading(false)
     setCategories([])
     setStatus("")
     setRevenueInput("")
@@ -221,35 +98,8 @@ export function useProjectImport({
     setTechnologies([])
     setActiveMetric(null)
     setGithubSearchQuery("")
-  }, [])
-
-  const captureScreenshot = useCallback(async (url: string, timeoutMs = 25000) => {
-    const controller = new AbortController()
-    const timer = window.setTimeout(() => controller.abort(), timeoutMs)
-
-    try {
-      const response = await fetch("/api/portfolio/screenshot", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ url }),
-        signal: controller.signal,
-      })
-
-      if (!response.ok) {
-        const errorBody = await response.json().catch(() => ({}))
-        devLog("Screenshot request failed", response.status, errorBody)
-        return null
-      }
-
-      const data = (await response.json()) as { screenshot?: string }
-      return data.screenshot || null
-    } catch (error) {
-      devLog("Screenshot request error", error)
-      return null
-    } finally {
-      window.clearTimeout(timer)
-    }
-  }, [])
+    resetMetadata()
+  }, [resetMetadata])
 
   useEffect(() => {
     if (!open) {
@@ -258,228 +108,32 @@ export function useProjectImport({
     }
 
     setIsFromGitHubSelection(false)
-  }, [open, resetForm])
+  }, [open, resetForm, setIsFromGitHubSelection])
 
-  useEffect(() => {
-    if (!open) return
+  const handleAddByGithub = useCallback(
+    (repo: RepositoryLike) => {
+      setTitle(repo.name || "")
 
-    const normalizedUrl = projectUrl.trim()
-    if (normalizedUrl.length < 8) return
-
-    if (isGitHubUrl(normalizedUrl)) {
-      if (isFromGitHubSelection) return
-
-      const parsed = parseGitHubRepoFromUrl(normalizedUrl)
-      if (!parsed) return
-
-      const githubImage = generateGitHubOgImage(parsed.owner, parsed.repoName)
-      setLogo(githubImage.proxyUrl)
-      setDirectLogoUrl(githubImage.directUrl)
-      setFavicon(null)
-      setIsFromGitHubSelection(false)
-      setImageLoadError(false)
-      setImageLoading(true)
-      return
-    }
-
-    const timer = window.setTimeout(async () => {
-      try {
-        setIsLoading(true)
-
-        const response = await fetch("/api/extract-metadata", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ url: normalizedUrl }),
-        })
-
-        if (!response.ok) {
-          console.error("extract-metadata API returned error:", response.status, await response.text())
-          return
-        }
-
-        const data = (await response.json()) as ExtractMetadataResponse
-        const projectData = data.projectData
-        const metadata = data.metadata
-
-        setTitle(projectData?.name || metadata?.title || "")
-        setDescription(projectData?.description || metadata?.description || "")
-        setFavicon(projectData?.favicon || metadata?.favicon || null)
-
-        setImageLoadError(false)
-        setImageLoading(true)
-
-        const metadataImage = getMetadataImage(metadata, projectData)
-        let nextDisplayLogo = metadataImage?.displayUrl ?? null
-        let nextDirectLogo = metadataImage?.directUrl ?? null
-
-        if (!nextDisplayLogo) {
-          const screenshot = await captureScreenshot(normalizedUrl)
-          if (screenshot) {
-            nextDisplayLogo = screenshot
-            nextDirectLogo = screenshot
-          }
-        }
-
-        setLogo(nextDisplayLogo)
-        setDirectLogoUrl(nextDirectLogo)
-
-        if (!nextDisplayLogo) {
-          setImageLoadError(true)
-          setImageLoading(false)
-        }
-      } catch (error) {
-        devLog("Metadata fetch failed", error)
-      } finally {
-        setIsLoading(false)
-      }
-    }, 500)
-
-    return () => window.clearTimeout(timer)
-  }, [open, projectUrl, isFromGitHubSelection, captureScreenshot])
-
-  const handleAddByGithub = useCallback((repo: RepositoryLike) => {
-    setTitle(repo.name || "")
-
-    if (repo.description && !GITHUB_DEFAULT_DESCRIPTION_PATTERN.test(repo.description.trim())) {
-      setDescription(repo.description)
-    } else {
-      setDescription("")
-    }
-
-    setFavicon(null)
-
-    const githubImage = getGitHubImageFromRepository(repo)
-    if (githubImage.proxyUrl) {
-      setLogo(githubImage.proxyUrl)
-      setDirectLogoUrl(githubImage.directUrl || githubImage.proxyUrl)
-      setIsFromGitHubSelection(true)
-      setImageLoadError(false)
-      setImageLoading(true)
-    } else {
-      setLogo(null)
-      setDirectLogoUrl(null)
-      setIsFromGitHubSelection(false)
-      setImageLoadError(false)
-      setImageLoading(false)
-    }
-
-    setProjectUrl(repo.htmlUrl || "")
-    setGithubSearchQuery("")
-
-    if (repo.languages && repo.languages.length > 0) {
-      setTechnologies(repo.languages)
-      return
-    }
-
-    if (repo.language) {
-      setTechnologies([repo.language])
-    }
-  }, [])
-
-  const handleUploadImage = useCallback(async (file: File) => {
-    const imageAsDataUri = await readFileAsDataUrl(file)
-    if (!imageAsDataUri) return
-
-    setLogo(imageAsDataUri)
-    setDirectLogoUrl(imageAsDataUri)
-    setImageLoadError(false)
-    setImageLoading(false)
-  }, [])
-
-  const handleImageLoadStart = useCallback(() => {
-    setImageLoading(true)
-  }, [])
-
-  const handleImageLoad = useCallback(
-    async (event: SyntheticEvent<HTMLImageElement>) => {
-      const target = event.currentTarget
-
-      if (logo && logo.includes("/api/portfolio/github-og-image")) {
-        try {
-          const response = await fetch(logo, { method: "HEAD" })
-          const actualOgImageUrl = response.headers.get("X-Actual-OG-Image-URL")
-          if (actualOgImageUrl) {
-            setDirectLogoUrl(actualOgImageUrl)
-          }
-        } catch {
-          // Keep current directLogoUrl when header fetch fails.
-        }
-
-        const width = target.naturalWidth
-        const height = target.naturalHeight
-
-        if (width > 0 && height > 0) {
-          const aspectRatio = width / height
-          const isSquare = Math.abs(aspectRatio - 1) < 0.15
-
-          if (isSquare && width >= 600) {
-            setImageLoadError(true)
-            setImageLoading(false)
-            return
-          }
-        }
+      if (repo.description && !GITHUB_DEFAULT_DESCRIPTION_PATTERN.test(repo.description.trim())) {
+        setDescription(repo.description)
+      } else {
+        setDescription("")
       }
 
-      setImageLoading(false)
-      setImageLoadError(false)
+      applyGitHubRepository(repo)
+      setProjectUrl(repo.htmlUrl || "")
+      setGithubSearchQuery("")
+
+      if (repo.languages && repo.languages.length > 0) {
+        setTechnologies(repo.languages)
+        return
+      }
+
+      if (repo.language) {
+        setTechnologies([repo.language])
+      }
     },
-    [logo]
-  )
-
-  const handleImageError = useCallback(
-    async (event: SyntheticEvent<HTMLImageElement>) => {
-      const target = event.currentTarget
-      const failedSrc = target.src
-      const normalizedProjectUrl = projectUrl.trim()
-      const canTryScreenshot = Boolean(normalizedProjectUrl) && !isGitHubUrl(normalizedProjectUrl)
-
-      if (failedSrc.includes("/api/portfolio/og-image-proxy") || failedSrc.includes("/api/portfolio/github-og-image")) {
-        try {
-          const response = await fetch(failedSrc)
-          if (!response.ok) {
-            if (response.status === 404 && canTryScreenshot) {
-              const screenshot = await captureScreenshot(normalizedProjectUrl, 20000)
-              if (screenshot) {
-                setLogo(screenshot)
-                setDirectLogoUrl(screenshot)
-                setImageLoadError(false)
-                setImageLoading(true)
-                return
-              }
-            }
-
-            setImageLoading(false)
-            setImageLoadError(true)
-            return
-          }
-
-          const blob = await response.blob()
-          target.src = URL.createObjectURL(blob)
-          setImageLoading(true)
-          return
-        } catch (error) {
-          devLog("Proxy image fetch failed", error)
-          setImageLoading(false)
-          setImageLoadError(true)
-          return
-        }
-      }
-
-      if (canTryScreenshot && !failedSrc.startsWith("data:")) {
-        const screenshot = await captureScreenshot(normalizedProjectUrl, 20000)
-        if (screenshot) {
-          setLogo(screenshot)
-          setDirectLogoUrl(screenshot)
-          setImageLoadError(false)
-          setImageLoading(true)
-          return
-        }
-      }
-
-      setImageLoading(false)
-      setImageLoadError(true)
-    },
-    [projectUrl, captureScreenshot]
+    [applyGitHubRepository]
   )
 
   const handleDone = useCallback(async () => {
@@ -529,35 +183,30 @@ export function useProjectImport({
   }, [
     title,
     projectUrl,
+    description,
     directLogoUrl,
     logo,
-    description,
+    captureScreenshot,
     favicon,
+    onAddImportedProject,
     categories,
     status,
     revenueInput,
     mrrInput,
     usersInput,
     technologies,
-    onAddImportedProject,
     onCaptureInsights,
     onOpenChange,
     resetForm,
-    captureScreenshot,
   ])
 
   return {
     projectUrl,
     setProjectUrl,
-    isLoading,
     title,
     setTitle,
     description,
     setDescription,
-    favicon,
-    logo,
-    imageLoadError,
-    imageLoading,
     categories,
     setCategories,
     status,
@@ -577,6 +226,11 @@ export function useProjectImport({
     filteredRepos,
     handleAddByGithub,
     handleDone,
+    isLoading,
+    favicon,
+    logo,
+    imageLoadError,
+    imageLoading,
     handleUploadImage,
     handleImageLoadStart,
     handleImageLoad,
